safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/document-write.html

<html>
<head>

<script>console.log("Line 4");</script>
<script>console.log("Line 5");</script>
<script>document.write("<scrip" + "t>console.log(\"Line 1 in document.write\");\n\n\nconsole.log(\"Line 4 in document.write\");</sc" + "ript>");</script>
<script>console.log("Line 7");</script>

<script>

function onload()
{
    if (window.testRunner)
        testRunner.dumpAsText();

    var script = document.createElement("script");
    script.textContent = "console.log(\"script element\")";
    document.head.appendChild(script);
    document.write("<scrip" + "t>console.log(\"document.write from onload\");</sc" + "ript>");
}

</script>
</head>

<body onload="console.log('Line 25'); onload();">
<p>
Tests that console reports zero line number for scripts generated with document.write. See https://bugs.webkit.org/show_bug.cgi?id=71099.
</p>

</body>
</html>
